refactor(#3272): extract wasi IO helper subsystem from index.ts (refs #3182)
WASI/node:fs import registration + deferred clock/fd_write/string-encode/
Uint8Array/ArrayBuffer/writeFileSync/sleep helpers moved verbatim. wasi.ts
imports LINEAR_U8_ARENA_START from linear-type-reservations; index.ts imports
back the 3 driver entry points + re-exports the 6 public helpers + 5 consts.
Byte-identity IDENTICAL (39/39), tsc 0.

✓
